+++ lib/septeo_design_system.dart
@@ -3,6 +3,8 @@ library;
 import 'package:flutter/material.dart';
 
 import 'foundations/main_theme.dart';
+
+export 'atomes/atomes_module.dart';
 export 'foundations/foundations_module.dart';
 
 class SepeteoDesignSystem {